Create logout component and add it to the drawer
